backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pymupdf==1.25.2
pydantic==2.10.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
requests==2.31.0
pytest==8.3.4
httpx==0.28.1